# crates/recall-server/src/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cortex-MCP HTTP API",
    "version": "0.1.0",
    "description": "Cortex local-first semantic memory broker. Hybrid retrieval (BM25 + cosine + recency decay) with explainable score breakdowns. All /v1/* endpoints require the API key when RECALL_MCP_API_KEY is set (Authorization: Bearer or X-API-Key)."
  },
  "paths": {
...
    "/v1/memories/{id}": {
      "delete": {
        "summary": "Forget a memory",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "204": { "description": "Deleted" },
          "404": { "$ref": "#/components/responses/Error" },
          "401": { "$ref": "#/components/responses/Error" }
        }
      },
      "patch": {
        "summary": "Update a memory in place (text is re-embedded automatically)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MemoryUpdate" } } }
        },
        "responses": {
          "200": { "description": "Updated memory", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Memory" } } } },
          "400": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" },
          "401": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/v1/recall": {
      "post": {
        "summary": "Hybrid recall with score breakdowns",
        "description": "total = w_k*norm_bm25 + w_v*cosine + w_r*exp(-days/tau) + pinned_boost",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["namespace", "query"],
                "properties": {
                  "namespace": { "type": "string" },
                  "query": { "type": "string", "minLength": 1 },
                  "k": { "type": "integer", "minimum": 1, "maximum": 100, "default": 5 },
                  "tau_days": { "type": ["number", "null"], "exclusiveMinimum": 0, "default": 30, "description": "null disables the recency term" },
                  "weights": {
                    "type": "object",
                    "properties": {
                      "bm25": { "type": "number", "default": 0.45 },
                      "vector": { "type": "number", "default": 0.45 },
                      "recency": { "type": "number", "default": 0.1 }
                    }
                  },
                  "pinned_boost": { "type": "number", "default": 0.2 },
                  "tags": {
                    "type": "array",
                    "items": { "type": "string" },
                    "default": [],
                    "description": "Only return memories carrying ALL of these tags (exact match)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ranked hits, each with its score breakdown",
            "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/RecallHit" } } } }
          },
          "400": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" },
          "401": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
  },
  "components": {
    "responses": {
      "Error": {
        "description": "Error",
        "content": { "application/json": { "schema": { "type": "object", "properties": { "error": { "type": "string" } } } } }
      },
...
    },
    "schemas": {
...
      "Memory": {
        "type": "object",
        "required": ["id", "namespace_id", "text", "tags", "source", "pinned", "created_at", "last_accessed_at"],
        "properties": {
          "id": { "type": "string" },
          "namespace_id": { "type": "string" },
          "text": { "type": "string" },
          "tags": { "type": "array", "items": { "type": "string" } },
          "embedding": { "type": ["array", "null"], "items": { "type": "number" } },
          "source": { "type": "string" },
          "pinned": { "type": "boolean" },
          "created_at": { "type": "integer" },
          "last_accessed_at": { "type": "integer" }
        }
      },
...
      "MemoryUpdate": {
        "type": "object",
        "description": "Partial update: omitted fields stay unchanged. A changed text is re-embedded by the server.",
        "properties": {
          "text": { "type": "string", "minLength": 1 },
          "tags": { "type": "array", "items": { "type": "string" } },
          "source": { "type": "string" },
          "pinned": { "type": "boolean" }
        }
      },
      "ScoreBreakdown": {
        "type": "object",
        "required": ["bm25", "vector", "recency", "pinned_boost", "total"],
        "properties": {
          "bm25": { "type": "number", "description": "normalized keyword score in [0,1]" },
          "vector": { "type": "number", "description": "cosine similarity in [-1,1]" },
          "recency": { "type": "number", "description": "exp(-days/tau) in (0,1]; 0 when decay disabled" },
          "pinned_boost": { "type": "number" },
          "total": { "type": "number" }
        }
      },
      "RecallHit": {
        "allOf": [
          { "$ref": "#/components/schemas/Memory" },
          {
            "type": "object",
            "required": ["breakdown"],
            "properties": { "breakdown": { "$ref": "#/components/schemas/ScoreBreakdown" } }
          }
        ]
      },
...
    }
  }
}